Description
This project describes the CO2 emission levels in the top 10 countries for the last 50 years. Python with imported libraries such as Seaborn and pandas was used to analyze the data. Plots were produced such as a histogram and a box plot to summarize the results of examining the inputs.
Python program was used to perform data analysis.
owid-co2-data.csv file belongs to owid repository (Our World in Data), https://github.com/owid/co2-data. The owid-co2-data.csv file was downloaded from Our World in Data on 03/29/2023.
loc function was significant in clearing the data; it narrowed down massive information to the required piece of information needed in the analysis as shown in Ibtisam MATH 2315 Lab 1. 
Author: Ibtisam Alkhawfi.
License: all material in this repository can be used by anyone.